+++ src/DotNetty.Common/Utilities/TaskEx.cs
@@ -35,5 +35,86 @@ namespace DotNetty.Common.Utilities
             tcs.TrySetException(exception);
             return tcs.Task;
         }
+
+        static readonly Action<Task, object> LinkOutcomeContinuationAction = (t, tcs) =>
+        {
+            switch (t.Status)
+            {
+                case TaskStatus.RanToCompletion:
+                    ((TaskCompletionSource)tcs).TryComplete();
+                    break;
+                case TaskStatus.Canceled:
+                    ((TaskCompletionSource)tcs).TrySetCanceled();
+                    break;
+                case TaskStatus.Faulted:
+                    ((TaskCompletionSource)tcs).TrySetException(t.Exception);
+                    break;
+                default:
+                    throw new ArgumentOutOfRangeException();
+            }
+        };
+
+        public static void LinkOutcome(this Task task, TaskCompletionSource taskCompletionSource)
+        {
+            switch (task.Status)
+            {
+                case TaskStatus.RanToCompletion:
+                    taskCompletionSource.TryComplete();
+                    break;
+                case TaskStatus.Canceled:
+                    taskCompletionSource.TrySetCanceled();
+                    break;
+                case TaskStatus.Faulted:
+                    taskCompletionSource.TrySetException(task.Exception);
+                    break;
+                default:
+                    task.ContinueWith(
+                        LinkOutcomeContinuationAction,
+                        taskCompletionSource,
+                        TaskContinuationOptions.ExecuteSynchronously);
+                    break;
+            }
+        }
+
+        class LinkOutcomeActionHost<T>
+        {
+            public static readonly Action<Task<T>, object> Action =
+                (t, tcs) =>
+                {
+                    switch (t.Status)
+                    {
+                        case TaskStatus.RanToCompletion:
+                            ((TaskCompletionSource<T>)tcs).TrySetResult(t.Result);
+                            break;
+                        case TaskStatus.Canceled:
+                            ((TaskCompletionSource<T>)tcs).TrySetCanceled();
+                            break;
+                        case TaskStatus.Faulted:
+                            ((TaskCompletionSource<T>)tcs).TrySetException(t.Exception);
+                            break;
+                        default:
+                            throw new ArgumentOutOfRangeException();
+                    }
+                };
+        }
+
+        public static void LinkOutcome<T>(this Task<T> task, TaskCompletionSource<T> taskCompletionSource)
+        {
+            switch (task.Status)
+            {
+                case TaskStatus.RanToCompletion:
+                    taskCompletionSource.TrySetResult(task.Result);
+                    break;
+                case TaskStatus.Canceled:
+                    taskCompletionSource.TrySetCanceled();
+                    break;
+                case TaskStatus.Faulted:
+                    taskCompletionSource.TrySetException(task.Exception);
+                    break;
+                default:
+                    task.ContinueWith(LinkOutcomeActionHost<T>.Action, taskCompletionSource, TaskContinuationOptions.ExecuteSynchronously);
+                    break;
+            }
+        }
     }
 }